+++ packages/react-intersection-observer/CHANGELOG.md
@@ -7,6 +7,8 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 <!-- ## [Unreleased] -->
 
+- Add `defaultLocale` option to Babel plugin.
+
 ## [2.0.8] - 2019-10-31
 
 - Change `useIntersection` hook behaviour to set and distribute the `IntersectionObserverEntry` object directly